# lxhcool-backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/public/projects/{slug}": {
      "get": {
        "operationId": "PublicProjectsController_getBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProjectResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Public Projects"
        ]
      }
    },
    "/admin/projects": {
      "get": {
        "operationId": "AdminProjectsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProjectResponseDto"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin Projects"
        ]
      },
      "post": {
        "operationId": "AdminProjectsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProjectResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin Projects"
        ]
      }
    },
    "/admin/projects/{id}": {
      "get": {
        "operationId": "AdminProjectsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProjectResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin Projects"
        ]
      },
      "patch": {
        "operationId": "AdminProjectsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProjectResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin Projects"
        ]
      },
      "delete": {
        "operationId": "AdminProjectsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProjectResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin Projects"
        ]
      }
    },
...
    "/admin/uploads/images": {
      "post": {
        "operationId": "UploadsController_uploadImage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/UploadResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Uploads"
        ]
      }
    },
    "/admin/uploads/audio": {
      "post": {
        "operationId": "UploadsController_uploadAudio",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/UploadResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Uploads"
        ]
      }
    },
    "/admin/uploads/videos": {
      "post": {
        "operationId": "UploadsController_uploadVideo",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/UploadResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Uploads"
        ]
      }
    },
...
    "/admin/widgets": {
      "get": {
        "operationId": "AdminWidgetsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/WidgetResponseDto"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin Widgets"
        ]
      },
      "post": {
        "operationId": "AdminWidgetsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWidgetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/WidgetResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin Widgets"
        ]
      }
    },
    "/admin/widgets/{id}": {
      "patch": {
        "operationId": "AdminWidgetsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWidgetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/WidgetResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin Widgets"
        ]
      },
      "delete": {
        "operationId": "AdminWidgetsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/WidgetResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin Widgets"
        ]
      }
    },
...
  },
  "info": {
    "title": "lxhcool API",
    "description": "API documentation for the lxhcool frontend and admin applications.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UploadResponseDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          }
        },
        "required": [
          "url"
        ]
      },
...
      "ProjectResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "excerpt": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "coverImage": {
            "type": "string",
            "nullable": true
          },
          "gallery": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "techStack": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "liveUrl": {
            "type": "string",
            "nullable": true
          },
          "githubUrl": {
            "type": "string",
            "nullable": true
          },
          "featured": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "HIDDEN"
            ]
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "seoTitle": {
            "type": "string",
            "nullable": true
          },
          "seoDescription": {
            "type": "string",
            "nullable": true
          },
          "ogImage": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "title",
          "slug",
          "excerpt",
          "content",
          "coverImage",
          "gallery",
          "techStack",
          "liveUrl",
          "githubUrl",
          "featured",
          "sortOrder",
          "status",
          "publishedAt",
          "seoTitle",
          "seoDescription",
          "ogImage",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "excerpt": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "coverImage": {
            "type": "string"
          },
          "gallery": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "techStack": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "liveUrl": {
            "type": "string"
          },
          "githubUrl": {
            "type": "string"
          },
          "featured": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "HIDDEN"
            ]
          },
          "seoTitle": {
            "type": "string"
          },
          "seoDescription": {
            "type": "string"
          },
          "ogImage": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "slug"
        ]
      },
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "excerpt": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "coverImage": {
            "type": "string"
          },
          "gallery": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "techStack": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "liveUrl": {
            "type": "string"
          },
          "githubUrl": {
            "type": "string"
          },
          "featured": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "HIDDEN"
            ]
          },
          "seoTitle": {
            "type": "string"
          },
          "seoDescription": {
            "type": "string"
          },
          "ogImage": {
            "type": "string"
          }
        }
      },
...
      "WidgetResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "area": {
            "type": "string",
            "enum": [
              "LEFT",
              "RIGHT"
            ]
          },
          "verticalPosition": {
            "type": "string",
            "enum": [
              "TOP",
              "BOTTOM"
            ]
          },
          "horizontalOffset": {
            "type": "number"
          },
          "verticalOffset": {
            "type": "number"
          },
          "rotation": {
            "type": "number"
          },
          "type": {
            "type": "string",
            "enum": [
              "MUSIC_PLAYER",
              "HITOKOTO",
              "FRIEND_LINKS",
              "DATE_CARD",
              "PHOTO_GALLERY",
              "PROJECT_TREE",
              "KEYBOARD"
            ]
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "enabled": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "number"
          },
          "config": {
            "type": "object",
            "additionalProperties": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "area",
          "verticalPosition",
          "horizontalOffset",
          "verticalOffset",
          "rotation",
          "type",
          "title",
          "enabled",
          "sortOrder",
          "config",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateWidgetDto": {
        "type": "object",
        "properties": {
          "area": {
            "type": "string",
            "enum": [
              "LEFT",
              "RIGHT"
            ]
          },
          "verticalPosition": {
            "type": "string",
            "enum": [
              "TOP",
              "BOTTOM"
            ]
          },
          "horizontalOffset": {
            "type": "number",
            "minimum": -1000,
            "maximum": 3000
          },
          "verticalOffset": {
            "type": "number",
            "minimum": -1000,
            "maximum": 3000
          },
          "rotation": {
            "type": "number",
            "minimum": -45,
            "maximum": 45
          },
          "type": {
            "type": "string",
            "enum": [
              "MUSIC_PLAYER",
              "HITOKOTO",
              "FRIEND_LINKS",
              "DATE_CARD",
              "PHOTO_GALLERY",
              "PROJECT_TREE",
              "KEYBOARD"
            ]
          },
          "title": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "number"
          },
          "config": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "area",
          "type"
        ]
      },
      "UpdateWidgetDto": {
        "type": "object",
        "properties": {
          "area": {
            "type": "string",
            "enum": [
              "LEFT",
              "RIGHT"
            ]
          },
          "verticalPosition": {
            "type": "string",
            "enum": [
              "TOP",
              "BOTTOM"
            ]
          },
          "horizontalOffset": {
            "type": "number",
            "minimum": -1000,
            "maximum": 3000
          },
          "verticalOffset": {
            "type": "number",
            "minimum": -1000,
            "maximum": 3000
          },
          "rotation": {
            "type": "number",
            "minimum": -45,
            "maximum": 45
          },
          "type": {
            "type": "string",
            "enum": [
              "MUSIC_PLAYER",
              "HITOKOTO",
              "FRIEND_LINKS",
              "DATE_CARD",
              "PHOTO_GALLERY",
              "PROJECT_TREE",
              "KEYBOARD"
            ]
          },
          "title": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "number"
          },
          "config": {
            "type": "object",
            "additionalProperties": true
          }
        }
      }
    }
  }
}